requirements.txt
Pyrogram
TgCrypto
requests
Pillow